Feat: Update Gemini API configuration to use environment variable for…

+++ src/main/java/org/example/audio_ecommerce/service/ProductViewService.java
@@ -25,7 +25,7 @@ public interface ProductViewService {
             String wardCode,
             BigDecimal minPrice,
             BigDecimal maxPrice,
-            Double minRating,
+            BigDecimal minRating,
             Pageable pageable
     );
